ibm-ups: Fix race condition with state manager
Fix a possible race condition with the phosphor-chassis-state-manager
application.

The ibm-ups application requires at least three matching reads from the
UPS device before the status bits are considered valid.  This is
intended to avoid transitory read issues that produce invalid data.

Previously, a UPS Device object was created on D-Bus with initial
property values indicating that the UPS was not present.  After three
matching reads, the properties were updated to contain the actual UPS
status.

This behavior could cause problems for the
phosphor-chassis-state-manager application.  It monitors the UPS D-Bus
properties in order to decide whether to perform an auto-power-restart
of the system.

If phosphor-chassis-state-manager read the D-Bus properties after the
UPS object was created but before three matching reads, it would
incorrectly assume no UPS was present.  As a result, it might perform an
auto-power-restart when it should not due to a low UPS battery.

This commit makes the following changes to fix this race condition:
* Modifies the ibm-ups systemd service file
  * Specifies that ibm-ups is a D-Bus application and that it has not
    finished starting until it claims the service/bus name.
  * Specifies that ibm-ups needs to be started before the
    phosphor-chassis-state-manager application.
* Does four reads from the UPS device when initially creating the D-Bus
  UPS object.  This should ensure it contains the actual UPS status.

Also made a minor change to mask off (ignore) modem status bits that are
not used by the ibm-ups application.  This avoids a potential problem
where changes in irrelevant bits could cause consecutive reads to be
considered non-matching and not valid.

Tested:
* Verified phosphor-chassis-state-manager does not start until after the
  ibm-ups application has claimed the D-Bus service/bus name.
* Verified ibm-ups application does not claim the D-Bus bus name until
  after it has polled the UPS device 4 times to obtain the actual UPS
  status.
* Verified that phosphor-chassis-state-manager receives forced
  PropertiesChanged signals if the ibm-ups application is restarted.
* Verified that PLDM receives forced PropertiesChanged signals if the
  ibm-ups application is restarted.
* Verified that extraneous modem bits are masked off (ignored) when
  reading the UPS status.

Signed-off-by: Shawn McCarney <shawnmm@us.ibm.com>
Change-Id: I3167cceae40826319571bdd2b08be47d12bcbdfe